Remove validChannelNames from the olm.package schema
Several concerns were raised with this field. It can be a maintenance
burden keeping it up-to-date with the set of channels defined in
each bundle of the package, and validation-related fields don't
belong in the core `olm.package` type.

While the usefulness of fail-safe validation is not in question,
it seems a better solution would be to introduce new top level
schemas for validation criteria. Tools concerned with validation
can consume them, and tools not concerned with validation can
ignore them.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>