Remove support for using individual declarative configs as input
With this commit, it is no longer valid to pass an individual
declarative config JSON file with the --database flag of serve.

Now, the acceptable declarative config input formats are:
- directory
- tar file

This commit also removes the --objects flag that was added
in a previous commit on this branch. Now objects are expected
to be found within the declarative config directory/tar,
rooted in the "objects" directory.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>